# Conflicts:
#	src/main/java/ar/edu/utn/frba/dds/server/Router.java
#	src/main/java/ar/edu/utn/frba/dds/services/seeders/SeederRoles.java

+++ src/test/java/ar/edu/utn/frba/dds/models/repositories/contribucion/CuidadoHeladerasRepositoryTest.java
@@ -84,7 +84,7 @@ class CuidadoHeladerasRepositoryTest {
 
   @Test
   void testMesesActivos() {
-    int mesesActiva = new CuidadoHeladerasRepository().getMesesActivosCumulativos(colaborador);
+    Long mesesActiva = new CuidadoHeladerasRepository().getMesesActivosCumulativos(colaborador);
     assertEquals(5, mesesActiva);
   }
 }
